+++ include/py68k_builtin.h
@@ -69,6 +69,11 @@ Py68Status py68_builtin_assign_add(Py68Runtime *runtime,
 Py68Status py68_builtin_assign_remove(Py68Runtime *runtime,
                                       Py68U16 argument_count,
                                       Py68Value *arguments, Py68Value *result);
+#if defined(PY68K_AMIGA)
+Py68Status py68_builtin_load_library(Py68Runtime *runtime,
+                                     Py68U16 argument_count,
+                                     Py68Value *arguments, Py68Value *result);
+#endif
 Py68Status py68_builtin_list(Py68Runtime *runtime, Py68U16 argument_count,
                              Py68Value *arguments, Py68Value *result);
 Py68Status py68_builtin_tuple(Py68Runtime *runtime, Py68U16 argument_count,